+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Features
 - Added settings which allow the user to override the dark and light foreground colors
   - _darkForegroundColor_ setting (defaults to `15202b`) which can override the dark foreground on what is calculated as a light background
   - _lightForegroundColor_ setting (defaults to `e7e7e7`) which can override the light foreground on what is calculated as a dark background
+- `Add recommended favorites` is now a command
+- Recommended favorites are now being saved when a new version is installed. Avoid using the same names.
 
 ## 1.3.1